float division instead of Int, trying to work out sending radius into poly

+++ spirolateral/PolygonView.swift
@@ -12,13 +12,15 @@ let strokeStyle = StrokeStyle(lineWidth:4)
 
 struct PolygonView : View {
     var sideCount:Int
+//    var radius:CGFloat
     var scale:CGFloat = 1.0
     var color:Color = Color.white
     var lineWidth:CGFloat = 16.0
     var body: some View {
         GeometryReader { geometry in
             Path { path in
-                let r = min(geometry.size.width, geometry.size.height) / 2
+                let r = CGFloat(min(geometry.size.width, geometry.size.height)) / 2.0
+//                let r = self.radius
                 let rScaled = r * self.scale
                 let points = Polygon.pointArray(sides: self.sideCount, cx: r, cy: r, radius: rScaled, offsetDegrees: 90)
                 path.move(
